+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.1.0
+
+* Adding sizes to icons
+* Updating the look and feel of modals
+* Add Breadcrumbs component
+* Add Sidebar component
+* Add Header component
+
 # 2.0.0
 
 * Rename library to @puppet/react-components